% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_client.R
\name{mcp_connect}
\alias{mcp_connect}
\title{Connect to V&A MCP Server}
\usage{
mcp_connect(url = "https://mcp.vam.ac.uk/mcp/")
}
\arguments{
\item{url}{Base URL of the MCP server.}
}
\value{
A V&A MCP client object (list).
}
\description{
Connect to V&A MCP Server
}
