% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_client.R
\name{mcp_extract}
\alias{mcp_extract}
\title{Extract Object Data}
\usage{
mcp_extract(client, id)
}
\arguments{
\item{client}{A vanda_client object.}

\item{id}{Object ID.}
}
\value{
A list containing object details.
}
\description{
Extract Object Data
}
