% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_client.R
\name{mcp_search}
\alias{mcp_search}
\title{Search V&A Collections}
\usage{
mcp_search(client, query)
}
\arguments{
\item{client}{A vanda_client object.}

\item{query}{Search query string.}
}
\value{
A data frame of results.
}
\description{
Search V&A Collections
}
